# packages/openapi/schema.json
{
  "components": {
    "schemas": {
...
      "AllergenGroup": {
        "description": "The groups elimination diets are built from.\n\nThe first six are the six-food elimination diet (6FED) groups, split where the\ndiet treats them as one \u2014 nuts into peanut and tree nut, seafood into fish and\nshellfish \u2014 so that a 4FED or 2FED is expressible without losing the finer\ngrain. Sesame is here because US labelling added it in 2023, not because the\nEoE diets eliminate it.\n\nPENDING CLINICAL CONFIRMATION: the group assigned to each catalog ingredient\n(migration 0004) needs the clinical advisor's review before it drives any\npatient-facing insight.",
        "enum": [
          "milk",
          "wheat",
          "egg",
          "soy",
          "peanut",
          "tree_nut",
          "fish",
          "shellfish",
          "sesame"
        ],
        "title": "AllergenGroup",
        "type": "string"
      },
...
      "CustomIngredientRead": {
        "properties": {
          "allergen_groups": {
            "items": {
              "$ref": "#/components/schemas/AllergenGroup"
            },
            "title": "Allergen Groups",
            "type": "array"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "allergen_groups"
        ],
        "title": "CustomIngredientRead",
        "type": "object"
      },
      "CustomIngredientUpdate": {
        "additionalProperties": false,
        "description": "Rename an ingredient or tag its allergen groups.\n\nRetagging applies to every past day it was eaten, deliberately: the tag\ndescribes what the ingredient is, and \"hot sauce turned out to contain soy\"\nis a correction to history, not a change from now on.",
        "properties": {
          "allergen_groups": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AllergenGroup"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Allergen Groups"
          },
          "name": {
            "anyOf": [
              {
                "maxLength": 120,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          }
        },
        "title": "CustomIngredientUpdate",
        "type": "object"
      },
...
      "EntryMethod": {
        "description": "Whether the entry was logged on the day it describes.\n\nResearch weights same-day entries more heavily than recalled ones, and this\nis derived server-side from the patient's own timezone rather than trusted\nfrom the client.",
        "enum": [
          "same_day",
          "backfill"
        ],
        "title": "EntryMethod",
        "type": "string"
      },
...
      "FoodDataSource": {
        "enum": [
          "open_food_facts",
          "usda_fdc"
        ],
        "title": "FoodDataSource",
        "type": "string"
      },
      "FoodItemInput": {
        "additionalProperties": false,
        "properties": {
          "eaten_on": {
            "format": "date",
            "title": "Eaten On",
            "type": "string"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientRef"
            },
            "maxItems": 40,
            "title": "Ingredients",
            "type": "array"
          },
          "meal": {
            "$ref": "#/components/schemas/Meal"
          },
          "name": {
            "maxLength": 120,
            "title": "Name",
            "type": "string"
          },
          "product": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductRef"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "eaten_on",
          "meal",
          "name"
        ],
        "title": "FoodItemInput",
        "type": "object"
      },
      "FoodItemList": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/FoodItemRead"
            },
            "title": "Items",
            "type": "array"
          },
          "range_end": {
            "format": "date",
            "title": "Range End",
            "type": "string"
          },
          "range_start": {
            "format": "date",
            "title": "Range Start",
            "type": "string"
          }
        },
        "required": [
          "range_start",
          "range_end",
          "items"
        ],
        "title": "FoodItemList",
        "type": "object"
      },
      "FoodItemRead": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "eaten_on": {
            "format": "date",
            "title": "Eaten On",
            "type": "string"
          },
          "entry_method": {
            "$ref": "#/components/schemas/EntryMethod"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientRead"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "meal": {
            "$ref": "#/components/schemas/Meal"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "id",
          "eaten_on",
          "meal",
          "name",
          "entry_method",
          "product",
          "ingredients",
          "created_at",
          "updated_at"
        ],
        "title": "FoodItemRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "Meal": {
        "enum": [
          "breakfast",
          "lunch",
          "dinner",
          "snack"
        ],
        "title": "Meal",
        "type": "string"
      },
...
      "ProductIngredientRead": {
        "properties": {
          "additive_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Additive Class"
          },
          "allergen_groups": {
            "items": {
              "$ref": "#/components/schemas/AllergenGroup"
            },
            "title": "Allergen Groups",
            "type": "array"
          },
          "depth": {
            "title": "Depth",
            "type": "integer"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "recognized": {
            "title": "Recognized",
            "type": "boolean"
          }
        },
        "required": [
          "key",
          "name",
          "depth",
          "recognized",
          "note",
          "allergen_groups",
          "additive_class"
        ],
        "title": "ProductIngredientRead",
        "type": "object"
      },
      "ProductRead": {
        "description": "A product's label, freshly looked up and not yet logged.",
        "properties": {
          "attribution": {
            "title": "Attribution",
            "type": "string"
          },
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "declared_allergens": {
            "items": {
              "$ref": "#/components/schemas/AllergenGroup"
            },
            "title": "Declared Allergens",
            "type": "array"
          },
          "inferred_allergens": {
            "items": {
              "$ref": "#/components/schemas/AllergenGroup"
            },
            "title": "Inferred Allergens",
            "type": "array"
          },
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/ProductIngredientRead"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "ingredients_complete": {
            "title": "Ingredients Complete",
            "type": "boolean"
          },
          "ingredients_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ingredients Text"
          },
          "may_contain": {
            "items": {
              "$ref": "#/components/schemas/AllergenGroup"
            },
            "title": "May Contain",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/FoodDataSource"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          },
          "source_updated_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Updated At"
          }
        },
        "required": [
          "source",
          "source_id",
          "barcode",
          "name",
          "brand",
          "ingredients_text",
          "ingredients",
          "ingredients_complete",
          "declared_allergens",
          "may_contain",
          "inferred_allergens",
          "source_updated_at",
          "attribution"
        ],
        "title": "ProductRead",
        "type": "object"
      },
...
      "ProductSummaryRead": {
        "properties": {
          "barcode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Barcode"
          },
          "brand": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Brand"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/FoodDataSource"
          },
          "source_id": {
            "title": "Source Id",
            "type": "string"
          }
        },
        "required": [
          "source",
          "source_id",
          "barcode",
          "name",
          "brand"
        ],
        "title": "ProductSummaryRead",
        "type": "object"
      },
      "RecentFood": {
        "description": "A food the patient has logged before, ready to log again in one tap.\n\nThe ingredients are those of the most recent time it was logged, which is the\nbest guess at how the patient makes it now.",
        "properties": {
          "ingredients": {
            "items": {
              "$ref": "#/components/schemas/IngredientRead"
            },
            "title": "Ingredients",
            "type": "array"
          },
          "last_eaten_on": {
            "format": "date",
            "title": "Last Eaten On",
            "type": "string"
          },
          "meal": {
            "$ref": "#/components/schemas/Meal"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "product": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProductSnapshotRead"
              },
              {
                "type": "null"
              }
            ]
          },
          "times_logged": {
            "title": "Times Logged",
            "type": "integer"
          }
        },
        "required": [
          "name",
          "meal",
          "product",
          "ingredients",
          "times_logged",
          "last_eaten_on"
        ],
        "title": "RecentFood",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Patient-owned EoE symptom tracking and clinical reporting. Personal health record; not a medical record system, and not medical advice.",
    "title": "eoehelp API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/foods/products/barcode/{barcode}": {
      "get": {
        "operationId": "product_by_barcode_api_v1_foods_products_barcode__barcode__get",
        "parameters": [
          {
            "in": "path",
            "name": "barcode",
            "required": true,
            "schema": {
              "pattern": "^\\d{8,14}$",
              "title": "Barcode",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Product By Barcode",
        "tags": [
          "food"
        ]
      }
    },
    "/api/v1/foods/products/search": {
      "get": {
        "operationId": "search_products_api_v1_foods_products_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": true,
            "schema": {
              "maxLength": 100,
              "minLength": 2,
              "title": "Q",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 40,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ProductSummaryRead"
                  },
                  "title": "Response Search Products Api V1 Foods Products Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Search Products",
        "tags": [
          "food"
        ]
      }
    },
    "/api/v1/foods/products/{source}/{source_id}": {
      "get": {
        "operationId": "product_detail_api_v1_foods_products__source___source_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "source",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/FoodDataSource"
            }
          },
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "maxLength": 64,
              "pattern": "^[0-9A-Za-z_-]+$",
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Product Detail",
        "tags": [
          "food"
        ]
      }
    },
...
    "/api/v1/me/foods": {
      "get": {
        "description": "Logged food between two days, inclusive. Both default to the patient's today.",
        "operationId": "list_foods_api_v1_me_foods_get",
        "parameters": [
          {
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "From"
            }
          },
          {
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodItemList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Foods",
        "tags": [
          "food"
        ]
      },
      "post": {
        "operationId": "log_food_api_v1_me_foods_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodItemInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Log Food",
        "tags": [
          "food"
        ]
      }
    },
...
    "/api/v1/me/foods/ingredients/{ingredient_id}": {
      "patch": {
        "operationId": "update_custom_ingredient_api_v1_me_foods_ingredients__ingredient_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "ingredient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ingredient Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CustomIngredientUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CustomIngredientRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Custom Ingredient",
        "tags": [
          "food"
        ]
      }
    },
    "/api/v1/me/foods/recent": {
      "get": {
        "operationId": "recent_foods_api_v1_me_foods_recent_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RecentFood"
                  },
                  "title": "Response Recent Foods Api V1 Me Foods Recent Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Recent Foods",
        "tags": [
          "food"
        ]
      }
    },
    "/api/v1/me/foods/{item_id}": {
      "delete": {
        "operationId": "delete_food_api_v1_me_foods__item_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Food",
        "tags": [
          "food"
        ]
      },
      "put": {
        "description": "Replace a logged food, ingredients included.",
        "operationId": "update_food_api_v1_me_foods__item_id__put",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FoodItemInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FoodItemRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Food",
        "tags": [
          "food"
        ]
      }
    },
...
  }
}